requirements.txt
asgiref==3.7.2
autopep8==2.0.2
Django==4.2.3
django-cors-headers==4.2.0
django-mathfilters==1.0.0
django-recaptcha==3.0.0
django-smtp-ssl==1.0
django-tinymce==3.6.1
djangorestframework==3.14.0
gunicorn==21.2.0
Pillow==10.0.0
pycodestyle==2.10.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
